Refactor session/current_user logic
We refactor the current_user logic (session logic for signed_in/out) to
encapsulate the logic in a single place. `Chatter.Session` now handles
all of signing in/out and has helper functions for whether or not
someone is signed_in. And `Chatter.User` encapsulates the data for the
username. This makes it flexible to be a database-backed module in the
future if desired.