Add chat histories
Currently, when a new user comes into a chat room, there is no history
there. We add such a history in this commit.

We decide to use processes to hold the history. Each room has a process
that holds the history, and they are supervised so they can be
automatically restarted.

We create the history process when we create a room. We also create
history processes for all the rooms that exist in the database when we
start the application.